requirements.txt
# Web Framework & Server
flask==3.1.3
werkzeug==3.1.7
gunicorn==23.0.0

# AI / Machine Learning (CPU version for Render)
tensorflow-cpu==2.17.0
numpy<2.0.0
pillow==11.1.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
line-bot-sdk==3.7.0